Moved MtxDataSource to io package #1401
