+++ gradoop-flink/src/test/java/org/gradoop/flink/io/impl/mtx/MtxDataSourceTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradoop.flink.model.impl.operators.layouting;
+package org.gradoop.flink.io.impl.mtx;
 
 import org.gradoop.common.model.impl.pojo.EPGMEdge;
 import org.gradoop.common.model.impl.pojo.EPGMVertex;
